fix(shared): extract isTelemetryDisabled to separate module for Node-free stub
Addresses PR feedback: the telemetry stub was re-exporting isTelemetryDisabled
from telemetry-client.ts, which imports @segment/analytics-node. This caused
the stub to pull in Node-only dependencies despite its purpose being a
client-safe no-op implementation.

Changes:
- Created packages/shared/src/telemetry/env-check.ts with only process.env checks
- Updated telemetry-client.ts to import isTelemetryDisabled from env-check
- Updated telemetry-stub.ts to import from env-check instead of telemetry-client
- Updated telemetry/index.ts to export isTelemetryDisabled from env-check
- Updated test imports to use the new module

Verified: dist/telemetry-stub.* contains no references to @segment/analytics-node
